10504 - moved the mobile filters button up to SearchPage.jsx; deleted unneeded stuff from SearchResults.jsx; still a small layout issue to solve, at mobile and tablet
