+++ src/_scss/layouts/search/search.scss
@@ -5,8 +5,8 @@
     @import "./header/header";
     .search-contents {
         @include grid-container;
-        margin-left: auto;
-        margin-right: auto;
+        //margin-left: auto;
+        //margin-right: auto;
         max-width: 100%;
         margin: 0 math.div($outer-gutter, 2);
         @include media($large-screen) {
